+++ db/migrate/20200304032705_create_haibanhoishoku_hormones.rb
@@ -3,11 +3,11 @@ class CreateHaibanhoishokuHormones < ActiveRecord::Migration[6.0]
     create_table :haibanhoishoku_hormones do |t|
       t.references :report, null: false, foreign_key: true
       t.integer :bt, null: false
-      t.integer :e2
-      t.integer :fsh
-      t.integer :lh
-      t.integer :p4
-      t.integer :hcg
+      t.float :e2
+      t.float :fsh
+      t.float :lh
+      t.float :p4
+      t.float :hcg
 
       t.timestamps
     end
